Move detailed contact view to PersonListPanel encapsulation
The original plan for the detailed contact view was such that the
detailed view and contact list view could share the space used
originally by just the list view, which in implementation, is the
PersonListPanel.

It is difficult to accomplish this with the detailed view being
encapsulated in a ContactFullViewPanel and the list view being
encapsulated in the PersonListPanel, due to the constraints presented by
JavaFX and the UI architecture.

To make it easier to achieve the desired UI effect, let's,
* abandon ContactFullViewPanel and move the detailed view to within
PersonListPanel's encapsulation
* rename PersonListPanel to a more accurate name in a future commit.
