+++ src/main/java/seedu/address/ui/PersonListPanel.java
@@ -2,6 +2,7 @@ package seedu.address.ui;
 
 import java.util.logging.Logger;
 
+import javafx.collections.ListChangeListener;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.ListCell;
@@ -18,15 +19,41 @@ public class PersonListPanel extends UiPart<Region> {
     private final Logger logger = LogsCenter.getLogger(PersonListPanel.class);
 
     @FXML
+    private ListView<Person> contactFullViewPanel;
+
+    @FXML
     private ListView<Person> personListView;
 
     /**
      * Creates a {@code PersonListPanel} with the given {@code ObservableList}.
      */
-    public PersonListPanel(ObservableList<Person> personList) {
+    public PersonListPanel(ObservableList<Person> personList, ObservableList<Person> contactFullView) {
         super(FXML);
+        contactFullViewPanel.setItems(contactFullView);
+        contactFullViewPanel.setCellFactory(listView -> new DetailedPersonCardCell());
+        contactFullView.addListener(new EmptyListener(contactFullViewPanel));
         personListView.setItems(personList);
         personListView.setCellFactory(listView -> new PersonListViewCell());
+        personList.addListener(new EmptyListener(personListView));
+    }
+
+    class EmptyListener implements ListChangeListener<Person> {
+        private ListView<Person> listView;
+
+        public EmptyListener(ListView<Person> listView) {
+            this.listView = listView;
+        }
+
+        @Override
+        public void onChanged(Change c) {
+            int len = c.getList().size();
+            logger.info("change detected, list size: " + len);
+            if (len == 0) {
+                listView.setPrefHeight(0);
+            } else {
+                listView.setPrefHeight(800);
+            }
+        }
     }
 
     /**
@@ -46,4 +73,17 @@ public class PersonListPanel extends UiPart<Region> {
         }
     }
 
+    class DetailedPersonCardCell extends ListCell<Person> {
+        @Override
+        protected void updateItem(Person person, boolean empty) {
+            super.updateItem(person, empty);
+
+            if (empty || person == null) {
+                setGraphic(null);
+                setText(null);
+            } else {
+                setGraphic(new DetailedPersonCard(person).getRoot());
+            }
+        }
+    }
 }
