+++ src/main/resources/view/DetailedPersonCard.fxml
@@ -4,14 +4,11 @@
 <?import javafx.scene.layout.FlowPane?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
-
-
-<?import javafx.scene.image.ImageView?>
 <?import javafx.scene.canvas.Canvas?>
 <?import javafx.scene.layout.Region?>
-<VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
+
+<VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" prefHeight="300">
     <HBox nodeOrientation="LEFT_TO_RIGHT">
-        <ImageView fx:id="images"/>
         <VBox>
             <HBox>
                 <Label fx:id="name"/>
