+++ src/main/resources/view/MainWindow.fxml
@@ -45,9 +45,6 @@
           </padding>
         </StackPane>
 
-        <StackPane fx:id="contactFullViewPlaceholder" minHeight="0.0" prefHeight="0.0" nodeOrientation="RIGHT_TO_LEFT"
-                   styleClass="pane-with-border" VBox.vgrow="ALWAYS"/>
-
         <VBox fx:id="personList" styleClass="pane-with-border" minWidth="340" prefWidth="340" VBox.vgrow="ALWAYS">
           <padding>
             <Insets top="10" right="10" bottom="10" left="10" />
