+++ src/main/resources/view/PersonListPanel.fxml
@@ -4,5 +4,6 @@
 <?import javafx.scene.layout.VBox?>
 
 <VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
-  <ListView fx:id="personListView" VBox.vgrow="ALWAYS" />
+  <ListView fx:id="contactFullViewPanel" prefHeight="0" VBox.vgrow="ALWAYS" />
+  <ListView fx:id="personListView" VBox.vgrow="SOMETIMES" />
 </VBox>
